+++ data.go
@@ -23,6 +23,7 @@ type DataPageInput struct {
 type DataPageOutput struct {
 	Version               string
 	Information           string
+	Software              string
 	Company               string
 	Alarms                string
 	MenuOverview          string
@@ -141,6 +142,7 @@ func data(writer http.ResponseWriter, request *http.Request, _ httprouter.Params
 		return dataWorkplaces[i].WorkplaceName < dataWorkplaces[j].WorkplaceName
 	})
 	data.Workplaces = dataWorkplaces
+	data.Software = cachedSoftwareName
 	data.Information = "INF: Page processed in " + time.Since(timer).String()
 	tmpl := template.Must(template.ParseFiles("./html/data.html"))
 	_ = tmpl.Execute(writer, data)