+++ main.go
@@ -5,12 +5,15 @@ import (
 	"github.com/kardianos/service"
 	"net/http"
 	"os"
+	"strconv"
+	"time"
 )
 
-const version = "2021.2.3.11"
+const version = "2021.2.3.14"
 const serviceName = "System WebService"
 const serviceDescription = "System web interface"
 
+//const config = "user=postgres password=pj79.. dbname=system host=ec2-3-17-5-15.us-east-2.compute.amazonaws.com port=5432 sslmode=disable application_name=system_webservice"
 //const config = "user=postgres password=pj79.. dbname=system host=localhost port=5432 sslmode=disable application_name=system_webservice"
 const config = "user=postgres password=pj79.. dbname=system host=database port=5432 sslmode=disable application_name=system_webservice"
 const (
@@ -42,6 +45,7 @@ type program struct{}
 
 func main() {
 	logInfo("SYSTEM", serviceName+" ["+version+"] starting...")
+	logInfo("MAIN", "© "+strconv.Itoa(time.Now().Year())+" Petr Jahoda")
 	serviceConfig := &service.Config{
 		Name:        serviceName,
 		DisplayName: serviceName,